# Practical fix for securedrop-client#1025 Faster sync for all source