@kushaldas 2020-06-22
If v2 API is further away time line wise, I think we should look at this one.