@cfm 2022-11-22
I've updated the title and description here to reflect that it addresses freedomofpress/securedrop-client#1025, not #1025 here.  I think we should revisit this in the broader conversation around when and how to consider the next version of the Journalist API.