@cfm 2024-12-05
Thanks for bringing this proposal to push in this direction.  Closing this draft as superseded by other planned work.